+++ plugins/org.preesm.model.scenario/src/org/preesm/model/scenario/check/FifoTypeChecker.java
@@ -8,10 +8,11 @@ import org.preesm.commons.logger.PreesmLogger;
 import org.preesm.model.pisdf.Fifo;
 import org.preesm.model.pisdf.PiGraph;
 import org.preesm.model.scenario.Scenario;
+import org.preesm.model.scenario.util.DefaultTypeSizes;
 
 /**
  * This class check if all fifos of the algorithm have a known data type size.
- * 
+ *
  * @author ahonorat
  */
 public class FifoTypeChecker {
@@ -22,7 +23,7 @@ public class FifoTypeChecker {
 
   /**
    * Get all the fifo types for which the size is not specified in the scenario.
-   * 
+   *
    * @param scenario
    *          To be checked.
    * @return A list of missing fifo type sizes. Empty list if no algorithm specified.
@@ -35,19 +36,32 @@ public class FifoTypeChecker {
     }
     for (final Fifo f : graph.getAllFifos()) {
       final String typeName = f.getType();
-      final Long typeSize = scenario.getSimulationInfo().getDataTypes().get(typeName);
-      if (typeSize == null) {
+      // Search for typeName in the Scenario
+
+      // If typeName is known in the Scenario
+      if (scenario.getSimulationInfo().getDataTypes().containsKey(typeName)) {
+        continue;
+      }
+
+      final long typeSize = DefaultTypeSizes.getInstance().getTypeSize(typeName);
+
+      if (typeSize != DefaultTypeSizes.UNKNOWN_TYPE) {
+        // If typeName matches a default known type
+        scenario.getSimulationInfo().getDataTypes().put(typeName, typeSize);
+        PreesmLogger.getLogger().info(() -> "A default size of " + typeSize + " bits was used for '" + typeName
+            + "' in fifo '" + f.getId() + "' for this Workflow execution.");
+      } else {
+        // If typeName is completely unknown
         result.add("'" + typeName + "'");
         PreesmLogger.getLogger().warning(() -> "Unknown type: " + typeName + ", in Fifo: " + f.getId());
       }
-
     }
     return result;
   }
 
   /**
    * Check if all the fifo types have a size defined in the scenario.
-   * 
+   *
    * @param scenario
    *          To be checked.
    * @throws PreesmRuntimeException
@@ -61,5 +75,4 @@ public class FifoTypeChecker {
           "Cannot find the size of the following fifo types: " + formattedMissingDataTypeSizes + ".");
     }
   }
-
 }