+++ plugins/org.preesm.ui.pisdf/src/org/preesm/ui/pisdf/features/OpenRefinementFeature.java
@@ -96,8 +96,8 @@ public class OpenRefinementFeature extends AbstractCustomFeature {
     if ((pes != null) && (pes.length == 1)) {
       final Object bo = getBusinessObjectForPictogramElement(pes[0]);
       RefinementContainer rc = null;
-      if (bo instanceof Delay) {
-        rc = ((Delay) bo).getActor();
+      if (bo instanceof final Delay delay) {
+        rc = delay.getActor();
       } else if (bo instanceof Actor || bo instanceof InitActor) {
         rc = (RefinementContainer) bo;
       }
@@ -126,8 +126,8 @@ public class OpenRefinementFeature extends AbstractCustomFeature {
     if ((pes != null) && (pes.length == 1)) {
       final Object bo = getBusinessObjectForPictogramElement(pes[0]);
       RefinementContainer rc = null;
-      if (bo instanceof Delay) {
-        rc = ((Delay) bo).getActor();
+      if (bo instanceof final Delay delay) {
+        rc = delay.getActor();
       } else if (bo instanceof Actor || bo instanceof InitActor) {
         rc = (RefinementContainer) bo;
       }
@@ -148,7 +148,7 @@ public class OpenRefinementFeature extends AbstractCustomFeature {
             final IResource diagResource = ResourcesPlugin.getWorkspace().getRoot().getFile(diagramFile);
 
             // Check if the diaram file exists
-            if (diagResource == null) {
+            if (!diagResource.exists()) {
               MessageDialog.openError(dw.getShell(), "Problem opening editor",
                   "No diagram file for " + refinementPath.lastSegment());
             } else {