+++ plugins/org.preesm.ui.scenario/src/org/preesm/ui/scenario/editor/simulation/SimulationPage.java
@@ -90,7 +90,6 @@ import org.eclipse.ui.forms.widgets.Section;
 import org.preesm.commons.logger.PreesmLogger;
 import org.preesm.model.pisdf.Fifo;
 import org.preesm.model.scenario.Scenario;
-import org.preesm.model.scenario.ScenarioConstants;
 import org.preesm.model.scenario.impl.DataTypeImpl;
 import org.preesm.model.scenario.util.DefaultTypeSizes;
 import org.preesm.model.slam.ComponentInstance;
@@ -109,6 +108,9 @@ public class SimulationPage extends ScenarioPage {
   private static final String DATA_TYPE_NAME_TITLE = Messages.getString("Simulation.DataTypes.typeColumn");
 
   private static final String[] DATA_TYPE_TABLE_TITLES = { DATA_TYPE_NAME_TITLE, DATA_TYPE_SIZE_TITLE };
+
+  private static final String COM_NODE = "comNode";
+  private static final String OPERATOR = "operator";
 
   /**
    * The listener interface for receiving comboBox events. The class that is interested in processing a comboBox event
@@ -141,14 +143,13 @@ public class SimulationPage extends ScenarioPage {
 
     @Override
     public void widgetSelected(final SelectionEvent e) {
-      if (e.getSource() instanceof Combo) {
-        final Combo combo = ((Combo) e.getSource());
+      if (e.getSource() instanceof final Combo combo) {
         final String item = combo.getItem(combo.getSelectionIndex());
 
         final ComponentInstance compInstance = scenario.getDesign().getComponentInstance(item);
-        if (this.type.equals("operator")) {
+        if (this.type.equals(OPERATOR)) {
           SimulationPage.this.scenario.getSimulationInfo().setMainOperator(compInstance);
-        } else if (this.type.equals("comNode")) {
+        } else if (this.type.equals(COM_NODE)) {
           SimulationPage.this.scenario.getSimulationInfo().setMainComNode(compInstance);
         }
       }
@@ -199,12 +200,12 @@ public class SimulationPage extends ScenarioPage {
       // Main operator chooser section
       createComboBoxSection(managedForm, Messages.getString("Simulation.mainOperator.title"),
           Messages.getString("Simulation.mainOperator.description"),
-          Messages.getString("Simulation.mainOperatorSelectionTooltip"), "operator");
+          Messages.getString("Simulation.mainOperatorSelectionTooltip"), OPERATOR);
 
       // Main medium chooser section
       createComboBoxSection(managedForm, Messages.getString("Simulation.mainMedium.title"),
           Messages.getString("Simulation.mainMedium.description"),
-          Messages.getString("Simulation.mainMediumSelectionTooltip"), "comNode");
+          Messages.getString("Simulation.mainMediumSelectionTooltip"), COM_NODE);
 
       // Text modification listener that updates the average data size
       final ModifyListener averageDataSizeListener = new ModifyListener() {
@@ -364,7 +365,7 @@ public class SimulationPage extends ScenarioPage {
     combo.setToolTipText(tooltip);
 
     final Design design = this.scenario.getDesign();
-    if (type.equals("operator")) {
+    if (type.equals(OPERATOR)) {
       for (final ComponentInstance opId : design.getOrderedOperatorComponentInstances()) {
         combo.add(opId.getInstanceName());
       }
@@ -373,7 +374,7 @@ public class SimulationPage extends ScenarioPage {
       if (mainOperator != null) {
         combo.select(combo.indexOf(mainOperator.getInstanceName()));
       }
-    } else if (type.equals("comNode")) {
+    } else if (type.equals(COM_NODE)) {
       for (final ComponentInstance nodeId : design.getCommunicationComponentInstances()) {
         combo.add(nodeId.getInstanceName());
       }
@@ -460,8 +461,7 @@ public class SimulationPage extends ScenarioPage {
     tableViewer.setCellModifier(new ICellModifier() {
       @Override
       public void modify(final Object element, final String property, final Object value) {
-        if (element instanceof TableItem) {
-          final TableItem ti = (TableItem) element;
+        if (element instanceof final TableItem ti) {
           final DataTypeImpl data = (DataTypeImpl) ti.getData();
           final long oldValue = data.getValue();
           try {
@@ -482,8 +482,8 @@ public class SimulationPage extends ScenarioPage {
 
       @Override
       public Object getValue(final Object element, final String property) {
-        if (element instanceof DataTypeImpl) {
-          return Long.toString(((DataTypeImpl) element).getValue());
+        if (element instanceof final DataTypeImpl delayTypeImpl) {
+          return Long.toString(delayTypeImpl.getValue());
         }
         return "";
       }
@@ -543,7 +543,7 @@ public class SimulationPage extends ScenarioPage {
             dialogTitle, dialogMessage, init, validator);
         if (dialog.open() == Window.OK) {
           SimulationPage.this.scenario.getSimulationInfo().getDataTypes().put(dialog.getValue().trim(),
-              (long) ScenarioConstants.DEFAULT_DATA_TYPE_SIZE.getValue());
+              DefaultTypeSizes.getInstance().getTypeSizeOrDefault(dialog.getValue().trim()));
           tableViewer.refresh();
           propertyChanged(SimulationPage.this, IEditorPart.PROP_DIRTY);
           gd.heightHint = Math.max(50,
@@ -583,7 +583,7 @@ public class SimulationPage extends ScenarioPage {
         for (final Fifo f : scenario.getAlgorithm().getAllFifos()) {
           final String typeName = f.getType();
           SimulationPage.this.scenario.getSimulationInfo().getDataTypes().put(typeName,
-              DefaultTypeSizes.getInstance().getDefaultTypeSize(typeName));
+              DefaultTypeSizes.getInstance().getTypeSizeOrDefault(typeName));
         }
 
         tableViewer.refresh();
@@ -703,12 +703,10 @@ public class SimulationPage extends ScenarioPage {
     treeviewer.setLabelProvider(new LabelProvider() {
       @Override
       public String getText(final Object element) {
-        if (element instanceof ComponentInstance) {
-          final ComponentInstance ci = (ComponentInstance) element;
+        if (element instanceof final ComponentInstance ci) {
           return ci.getInstanceName() + " (" + ci.getComponent().getVlnv().getName() + ")";
-        } else {
-          return super.getText(element);
         }
+        return super.getText(element);
       }
     });
     treeviewer.setAutoExpandLevel(AbstractTreeViewer.ALL_LEVELS);