+++ tests/org.preesm.algorithm.tests/src/org/ietr/dftools/algorithm/test/GMLDAGImporterTest.java
@@ -80,12 +80,12 @@ public class GMLDAGImporterTest {
     } else {
       createTempFile = Files.createTempFile("export_test_", ".xml").toFile();
 
-      boolean fail = true;
+      boolean success = true;
 
-      fail &= createTempFile.setReadable(true, true);
-      fail &= createTempFile.setWritable(true, true);
-      fail &= createTempFile.setExecutable(true, true);
-      Assert.assertTrue(fail);
+      success &= createTempFile.setReadable(true, true);
+      success &= createTempFile.setWritable(true, true);
+      success &= createTempFile.setExecutable(true, true);
+      Assert.assertTrue(success);
     }
     createTempFile.deleteOnExit();