+++ tests/org.preesm.model.tests/src/org/preesm/tests/model/scenario/types/DefaultTypeSizesTest.java
@@ -34,9 +34,15 @@
  */
 package org.preesm.tests.model.scenario.types;
 
-import static org.junit.Assert.assertEquals;
-
+import java.util.Arrays;
+import java.util.Collection;
+import org.junit.Assert;
 import org.junit.Test;
+import org.junit.experimental.runners.Enclosed;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+import org.junit.runners.Parameterized.Parameters;
+import org.preesm.model.scenario.ScenarioConstants;
 import org.preesm.model.scenario.util.DefaultTypeSizes;
 
 /**
@@ -44,42 +50,73 @@ import org.preesm.model.scenario.util.DefaultTypeSizes;
  * @author anmorvan
  *
  */
+@RunWith(Enclosed.class)
 public class DefaultTypeSizesTest {
 
-  @Test
-  public void testVoidTypeSize() {
-    final long defaultTypeSize = DefaultTypeSizes.getInstance().getDefaultTypeSize("void");
-    assertEquals(1L, defaultTypeSize);
+  private DefaultTypeSizesTest() {
+
   }
 
-  @Test
-  public void testCharTypeSize() {
-    final long defaultTypeSize = DefaultTypeSizes.getInstance().getDefaultTypeSize("char");
-    assertEquals(8L, defaultTypeSize);
+  @RunWith(Parameterized.class)
+  public static class DefaultTypeSizesParamTest {
+    private final String typeName;
+    private final long   expectedSize;
+
+    public DefaultTypeSizesParamTest(String typeName, long expectedSize) {
+      this.typeName = typeName;
+      this.expectedSize = expectedSize;
+    }
+
+    @Parameters(name = "Test {1} size = {0}")
+    public static Collection<Object[]> data() {
+      return Arrays.asList(new Object[][] { { "DEFAULT", ScenarioConstants.DEFAULT_DATA_TYPE_SIZE_VALUE },
+          { "void", 1 }, { "char", 8 }, { "unsigned char", 8 }, { "short", 16 }, { "unsigned short", 16 },
+          { "int", 32 }, { "unsigned int", 32 }, { "long", 64 }, { "unsigned long", 64 }, { "long long", 64 },
+          { "unsigned long long", 64 }, { "float", 32 }, { "double", 64 }, { "long double", 128 }, { "int8_t", 8 },
+          { "int16_t", 16 }, { "int32_t", 32 }, { "int64_t", 64 }, { "int_fast8_t", 8 }, { "int_fast16_t", 64 },
+          { "int_fast32_t", 64 }, { "int_fast64_t", 64 }, { "int_least8_t", 8 }, { "int_least16_t", 16 },
+          { "int_least32_t", 32 }, { "int_least64_t", 64 }, { "intmax_t", 64 }, { "intptr_t", 64 }, { "uint8_t", 8 },
+          { "uint16_t", 16 }, { "uint32_t", 32 }, { "uint64_t", 64 }, { "uint_fast8_t", 8 }, { "uint_fast16_t", 64 },
+          { "uint_fast32_t", 64 }, { "uint_fast64_t", 64 }, { "uint_least8_t", 8 }, { "uint_least16_t", 16 },
+          { "uint_least32_t", 32 }, { "uint_least64_t", 64 }, { "uintmax_t", 64 }, { "uintptr_t", 64 } });
+    }
+
+    @Test
+    public void testTypeSizeParameterized() {
+      Assert.assertEquals(expectedSize, DefaultTypeSizes.getInstance().getTypeSizeOrDefault(typeName));
+    }
   }
 
-  @Test
-  public void testShortTypeSize() {
-    final long defaultTypeSize = DefaultTypeSizes.getInstance().getDefaultTypeSize("short");
-    assertEquals(16L, defaultTypeSize);
-  }
+  public static class DefaultTypeSizesSingleTest {
 
-  @Test
-  public void testINT32TTypeSize() {
-    final long defaultTypeSize = DefaultTypeSizes.getInstance().getDefaultTypeSize("int32_t");
-    assertEquals(32L, defaultTypeSize);
-  }
+    @Test
+    public void testUnknownTypeSize() {
+      final long unknownTypeSize = DefaultTypeSizes.getInstance().getTypeSize("THIS_IS_AN_UNKNOWN_TYPENAME");
+      Assert.assertEquals(DefaultTypeSizes.UNKNOWN_TYPE, unknownTypeSize);
+    }
 
-  @Test
-  public void testDoubleTypeSize() {
-    final long defaultTypeSize = DefaultTypeSizes.getInstance().getDefaultTypeSize("double");
-    assertEquals(64L, defaultTypeSize);
-  }
+    @Test
+    public void testVitisApIntSize() {
+      Assert.assertEquals(10L, DefaultTypeSizes.getInstance().getTypeSize("ap_int<10>"));
+      Assert.assertEquals(10L, DefaultTypeSizes.getInstance().getTypeSize("ap_uint<10>"));
 
-  @Test
-  public void testLongDoubleTypeSize() {
-    final long defaultTypeSize = DefaultTypeSizes.getInstance().getDefaultTypeSize("long double");
-    assertEquals(128L, defaultTypeSize);
-  }
+      Assert.assertEquals(DefaultTypeSizes.UNKNOWN_TYPE, DefaultTypeSizes.getInstance().getTypeSize("ap_int<0>"));
+      Assert.assertEquals(DefaultTypeSizes.UNKNOWN_TYPE, DefaultTypeSizes.getInstance().getTypeSize("ap_int<-1>"));
+    }
 
+    @Test
+    public void testVitisApFixedSize() {
+      Assert.assertEquals(10L, DefaultTypeSizes.getInstance().getTypeSize("ap_fixed<10,5>"));
+      Assert.assertEquals(10L, DefaultTypeSizes.getInstance().getTypeSize("ap_ufixed<10,5>"));
+      Assert.assertEquals(10L, DefaultTypeSizes.getInstance().getTypeSize("ap_ufixed<10,8, AP_RND>"));
+      Assert.assertEquals(10L, DefaultTypeSizes.getInstance().getTypeSize("ap_ufixed<10,8, AP_RND, AP_SAT>"));
+
+      Assert.assertEquals(DefaultTypeSizes.UNKNOWN_TYPE,
+          DefaultTypeSizes.getInstance().getTypeSize("ap_ufixed<0,8, AP_RND>"));
+      Assert.assertEquals(DefaultTypeSizes.UNKNOWN_TYPE,
+          DefaultTypeSizes.getInstance().getTypeSize("ap_ufixed<-1,8, AP_RND>"));
+      Assert.assertEquals(DefaultTypeSizes.UNKNOWN_TYPE,
+          DefaultTypeSizes.getInstance().getTypeSize("ap_ufixed<10,8, ERROR>"));
+    }
+  }
 }